requirements.txt
albumentations==1.1.0
image_slicer==2.1.1
matplotlib==3.5.0
natsort==8.0.0
numpy==1.21.4
opencv_python==4.5.4.60
pandas==1.3.4
Pillow==8.4.0
scikit_learn==1.0.1
setuptools==59.2.0
timm==0.4.12
torch==1.10.0
torchinfo==1.5.3
tqdm==4.62.3